rename: image 파일 경로 수정
